(Refactor) Reduce duplicated code in the CSV export services
All of the existing Export service classes use the same code. Reduce the
duplication by adding a superclass (non-initializable) to hold the duplicate
code.
